Support flycast db urls
We launched flycast for pg clusters earlier this year: https://community.fly.io/t/flycast-for-postgres/10628

Now DATABASE_URLs that use flycast will have the form: `postgres://user:pass@app-name.flycast:5432/db_name`. Before this change the host was rewritten to `top1.nearest.of.app-name.flycast`, which is not valid and would result in failed connections.

With this change we do not modify hostname when a flycast hostname is used.